Policy JSON saves in localStorage
The policy JSON will be saved in localStorage, and first then it on table and completly done they will get send over mqtt. The json data is getting saved in policies state in dragndropPolicies component. Rememeber they are string and needs to be JSON.parse before being used as json