frontpage data and config page data, emitter use removed
The emitter is no longer used for sending data from dragNdrop to table. MQTT broker used to send the data to the applied policies, and used "to quarry" that then fetches the data from localstorage needed on frontpage. Deleting of applied policies is not yet implemented and at this time only remove last element of a state.